Implement submit -> modalUnvisible
